Add files via upload
This code determines a monetary response based on a user's greeting. Here's how it works:

The user is prompted to input a greeting.
If the greeting starts with "hello" (case-insensitive), it prints $0.
If the greeting starts with "h" but not "hello" (case-insensitive), it prints $20.
For any other greeting, it prints $100.